New structure for `Domain` class
`Domain` class modified based to get SDF and BC at initiation.
`CookDomain` modified according to new class structure for illustration,
other example domains should be altered likewise.
The mesh generation process is explained in `test1` and usage
of example domains is illustrated with an example of `CookDomain`.
`DistFnc` does not need BdBox as an argument, so it is corrected all
over the code.